+++ lapdog/vue/src/Pages/Submission.vue
@@ -205,8 +205,8 @@ Other: error_outline
       </div>
     </div>
     <h4>
-      <router-link v-if="submission && submission.AUTHORIZED_DOMAIN" :to="{name: 'workspace', params: {namespace: lodash.split(submission.AUTHORIZED_DOMAIN)[0], workspace: lodash.split(submission.AUTHORIZED_DOMAIN)[1]}}">
-        <i class="material-icons tiny">arrow_back</i>  <span class="icon-align-header">{{lodash.split(submission.AUTHORIZED_DOMAIN)[0]}}/{{lodash.split(submission.AUTHORIZED_DOMAIN)[1]}}</span>
+      <router-link v-if="submission && submission.AUTHORIZED_DOMAIN" :to="{name: 'workspace', params: {namespace: namespace, workspace: lodash.split(submission.AUTHORIZED_DOMAIN, '/')[1]}}">
+        <i class="material-icons tiny">arrow_back</i>  <span class="icon-align-header">{{namespace}}/{{lodash.split(submission.AUTHORIZED_DOMAIN, '/')[1]}}</span>
       </router-link>
       <router-link v-else :to="{name: 'workspace', params: {namespace: namespace, workspace: workspace}}">
         <i class="material-icons tiny">arrow_back</i>  <span class="icon-align-header">{{namespace}}/{{workspace}}</span>